journal: copy component list when iterating to delete
This adds a rut_entity_foreach_component_safe function that can be used
to iterate the components while possibly altering the list of components
by taking a copy first. This is used by
rig_undo_journal_delete_entity_and_log otherwise we are left with
partially deleted entities and the inconsistent state can cause Rig to
crash when saving for example.