texture: fix texture cache
Although we were looking up in a hash table for existing textures
according to a given filename, we weren't actually adding cache
entries! This changes the hash table to use g_direct_equal/hash and
create a quark from the filename to use as a key for adding a record of
each texture loaded to the hash table.