engine: allow materials with model components
We used to disallow material components in conjunction with model
components in case the model had no texture coordinates because it was
too easy to lose a model by associating a texture with a transparent
top-left texel that would be mapped across the whole model. Since we
now use cylindrical texture mapping to generate texture coordinates if
the model doesn't have its own this shouldn't be a problem anymore.