Enable  select/copy/paste for asset inspectors
This makes asset inspectors selectable and adds visual feedback when an
asset inspector is selected. This also make assets mimable so they can
be copy and pasted.

Note: this change highlights a limitation with how renderers can track
changes to components. Currently we rely on rig-engine.c to notify the
renderer when we make changes to the image source of a material
component, but an asset inspector is not currently able to notify the
renderer and so updates via an asset inspector aren't reflected in
the UI preview. This requires some re-working of the renderer to fix
properly.