model: don't automatically generate patch/fin meshes
Since we only need patch/fin geometry when using the hair component this
updates the model code so we have a rut_model_new_for_hair() api that
can be used to derive a model from a previous model that is suitable for
hair rendering. This way we can make rig-engine.c trigger this code path
only when we adding a RutHair component to an entity.