hair: optimize how we track shells/particles
This replaces the use of GQueues with GArrays instead. In glib GQueues
are implemented as a linked list and the pop/push pattern being used to
iterate the queues will internally be unlinking/linking the item from
the list and freeing/allocating link nodes each time which isn't ideal
for repeated iteration of this data.