Adds RutDragBin to support drag initiation
This adds a convenient way to enable drag & drop support by adding a new
bin widget with an associated child and payload object. If it sees a
button press that the child itself hasn't handled then it will start a
grab and visually show that a drag has started. During the drag an icon
will follow the mouse and widgets under the cursor will receive
_DROP_OFFER events that give them the opportunity to claim the payload
when the mouse is released.

To verify the interface we now support dragging and dropping assets from
search results onto asset-inspectors.