MSVC Projects: Add project templates for cogl-path
This adds template project files that is used to build cogl-path, where the
source files can be filled in using autotools. Note that although cogl-path
is built as a static library, items to build DLLs for this is intentionally
left in there as this is intended to be a shared library later.