shim: allow only shimming a single axis
A RutShim is a way of forcibly taking up a fixed space and also forcing
a child to fit in that space. This allows us to control which axis will
have a fixed size so we can allow the un-constrained axis to be
negotiated via the get_preferred_/width/height apis.

By default both the x and y axis will be fixed