text: remove input region if non-selectable/editable
The remove the input_region of a RutText from the scenegraph if we know
it's not editable or selectable so that it doesn't inadvertently block
other input events. Even though the handlers would report that the
event was _UNHANDLED we would then traverse up the scenegraph for
fallback handlers which can change where the event would otherwise be
delivered.