object: Adds utilities for allocating objects
This adds _rut_object_alloc[0] functions for slice allocating an object
and calling rut_object_init(). This api will also verify that a given
RutType has been initialized and if not will call a given function to
initialize the type first. This will let us remove a good amount of of
boilerplate from the many objects we define in Rut/Rig.