Remove padding from eulers and quaternions
One of the notable advantages of a Euler representation of a rotation is
that it is compact but we negate that benefit if all CoglEulers have
enough padding to potentially cache a quaternion in the future. Having
padding for all CoglQuaternions is probably also unnecessary as higher
level code can find its own ways of caching data relating to
quaternions.

Reviewed-by: Neil Roberts <neil@linux.intel.com>