refable: Adds rut_refable_claim/_release apis
This adds new apis for taking references on objects in terms of a
specific object claiming a reference to another object. By explicitly
stating the ownership of a reference we can improve our reference count
debugging so when an object still has references that are keeping it
from being freed we can see the actual object that owns those
references. This makes it much easier to understand where leaks are
originating from.

In the future we should be able to also further improve our debugging
utilities to automatically traverse this graph of relationships and
whenever we find a sub-graph that doesn't terminate at a known top-level
object such as a RutContext or RutShell then we can immediately report
that a leak as occurred.