Fix cylindrical texture mapping
When integrating the fur rendering code I broke how we measure meshes
and how we generate missing attributes in the non fur case.

This adds some code for making a deep copy of a RutMesh so that we can
consistently avoid modifying an original asset mesh. Later we also need
to add support for serializing a mesh when we save a UI so we don't need
to derive anything at runtime.