Adds some ground-work for a selection tool
This adds a RigSelectionTool that shows a few control points while an
entity is selected. At the moment the control points can't be
manipulated but the plan is to use them to be able to control the width
and height or scale of entities/components. The control points are
drawn in an overlay, over the camera-view using the same scenegraph as
the 2D ui widgets.