prop-inspector: grey + block input to readonly properties
To explicitly stop the user from being able to change read-only
properties we now put the inspectors into a stack where we can add an
input-region over the top of the inspector widget that blocks any input
from reaching the widget itself. Similarly we can add a transparent grey
rectangle over the widget to show that it is disabled.