build: fix build with cogl master f75369854
In Cogl master the cogl-sdl.h header is now a top-level header that has
to be included explicitly and also since SDL.h must have had the side
effect of including <string.h> and <stdlib.h> we need to include those
in a couple of places to fix the build.

Cogl master also no longer provides type cast macros like COGL_TEXTURE()
so this patch removes all use of those.