Update marking of editor mode code
This renames _rig_in_device_mode to _rig_in_editor_mode since almost
every test of _rig_in_device_mode is negated assuming that if we aren't
in device mode then we must be in editor mode, which might not make
sense if we introduce other, non mutually exclusive mode/option states.
E.g what if we defined a "slave" mode for running rig-slave or if we
introduce a "simulator"/"headless" mode for updating properties without
graphics then we also wouldn't be in device mode but that wouldn't imply
we are running the editor either.