pb-rpc: peer to peer support
This adds support for peer to peer, protocol buffer based rpc
connections where each end may act as a client and a server while
sharing a single file descriptor.

Since the status field for client replies was actually being ignored
this removes the field so the headers for server requests and client
replies are the same. Replies can be distinguished from requests because
the method_index will == ~0 for replies.

A new PB_RPC_Peer object has been introduced that internally wraps a
PB_RPC_Client and a PB_RPC_Server which both share a newly defined
Stream object that handles buffered IO.

For now the code only supports peer to peer connections where each peer
already has an open file descriptor since our use case is that we will
be fork()ing a child process after using socketpair() to open two file
descriptors for a stream socket to be used for the peer to peer rpc
connection.