Run a headless simulator process
This lays the groundwork for moving UI logic into a separate process
from the rendering process.

When initializing the Rig engine in rig_engine_init we now fork and exec
a rig-simulator binary which we can communicate with via a peer to peer
rpc connection over a socket stream.

The intention is that the main renderer process will feed the simulator
process with any necessary input events for each frame, the simulator
will asynchronously update the graph of entities, by running all UI
logic and updating properties and for each frame it will send back the
changes to the renderer process.

This design will help us to avoid blocking the renderer process to keep
animations smooth and will also enable us to remote debug code in the
simulator so we can set breakpoints in the editor and step through code.