defer redraw control to rig programs
Instead of having a single _rut_shell_paint that controls the redraw
cycle each frame, we now expose several shell functions that can be used
to implement redraw functions in each of the rig binaries. This lets
us have a redraw function in the simulator process that doesn't call
rig_engine_paint() and will be able to send its property updates to the
renderer process at the end of each frame.