device: Forward input events to the simulator
Whenever we start a new frame in rig-device we request the simulator to
run a frame too. As input to each simulator frame we now forward all
input events that are currently queued up.

This introduces a new RutStreamEvent that can be considered to be the
native event for the simulator process. The native events received via
SDL in the renderer process are serialized to be sent to the simulator
process to become RutStreamEvents.