peer: associate user_data with server connection
This makes sure we associate the user_data given to rig_rpc_peer_new()
with the server-connection so that when a request is made to the server
we can get the user data via rig_pb_rpc_closure_get_connection_data().

This fixes a crash in the simulator when being requested to run a new
frame.