Support serializing and loading the UI in the simulator
This bootstraps being able to initialize the state of the simulator
process by being able to send it a description of the UI similar to
how we send the UI to a slave process.

Notably this adds support for serializing and unserializing RutMesh
objects and their associated attributes and buffers, instead of relying
on simply sending ply files over the wire; something we've needed for
some time really.

This exposes RigPBSerializer and RigPBUnSerializer objects for the state
tracking required for a single serialization job that may itself be
comprised of multiple actions and may need configuring depending on the
specific context in which we are serializing data.

These changes have also been highlighting the confusing role or RutAsset
objects. Assets let us track the original file path associated with some
data that was imported into Rig, as well as a thumbnail and user tags.
Assets are getting used as a general container for textures and meshes
though which is especially awkward in the context of the simulator
process where we don't need thumbnails and it's also really bad that we
repeatedly derive mesh data from assets instead of storing and
serializing the final derived data.