simulator: Add support for logging + serializing property changes
This starts adding support for logging all property changes using a
RutMemoryStack in the RutPropertyContext which can be used for
synchronizing property changes made in the Rig simulator process with
the frontend process each frame.

At the end of rig_simulator_run_frame() we now serialize all changed
properties and send them to the frontend process.