Support new github upstream for protobuf-c
The protobuf-c maintainers changed recently and it is now hosted on
github. This updates Rig to build with the latest version on github.
Since the upsream now better differentiates the protobuf-c library from
the protobuf-c-rpc code and since we already had our own copy of most of
the rpc related code under rig/jni/protobuf-c this takes a copy of the
remaining bits (the dispatch apis) so that we don't need to depend on
building the rpc code from upstream.