Add a RutMagazine utility api
This adds a simple, but very fast memory allocation api for allocating
objects of a pre-determined size.

This allocator builds on RutMemoryStack for making all initial
allocations but never frees memory back to the stack.

Memory chunks that haven't been allocated yet are stored in a singly
linked, fifo, list.

Allocating from a magazine is simply a question of popping an entry from
the head of the fifo list. If no entries are available then instead
allocate from the memory stack instead.

When an entry is freed, it is put back into the fifo list for re-use.

No attempt is ever made to shrink the amount of memory associated with a
RutMagazine.