start reading property changes in frontend
Adds initial code for sending property changes from the simulator to the
frontend and iterating through those changes in the frontend. Currently
though we don't have a way of looking up object properties using a
property id though, we have to use a name so we'll need to update the
introspectable interface a bit for this.