major overhaul of RutType and RutObject
The renames the widespread use of the term "interface" in Rut with
"trait". This is inspired by the Rust language whose trait model seems
similar to our mix-in model. The term interface was starting to seem
misleading.

This adds the ability for each trait of an object to optionally have a
corresponding destructor. This is because it's easier to forget to clean
up state for a trait in the _free function which it's disconnected from
the code that adds the trait. It's also just more convenient.

This adds the ability to associate a type with a RutMagazine for faster
allocation performance and also better data locality for objects of the
same type.

This gets rid of the refable interface and moves the ref_count into the
base object. This removes quite a bit of unnecessary indirection and
reduces the boilerplate needed when creating new types.

This removes the need for _rut_init() since all types are now
initialized lazily in the same way.